# Fix benchmark issues and update libraries