--- a/NetCore/NetCore.csproj
+++ b/NetCore/NetCore.csproj
@@ -1,12 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFrameworks>netcoreapp2.1;netcoreapp3.0</TargetFrameworks>
     <DebugType>portable</DebugType>
     <AssemblyName>NetCore</AssemblyName>
     <OutputType>Exe</OutputType>
     <Optimize>true</Optimize>
-    <PackageId>NetCore</PackageId>
     <RootNamespace>ImageProcessing</RootNamespace>
     <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
@@ -16,17 +15,17 @@
 
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.11.5" />
-    <PackageReference Include="FreeImage.Standard" Version="4.3.7" />
-    <PackageReference Include="Magick.NET-Q8-AnyCPU" Version="7.12.0" />
-    <PackageReference Include="NetVips" Version="1.1.0-rc1" />
+    <PackageReference Include="FreeImage.Standard" Version="4.3.8" />
+    <PackageReference Include="Magick.NET-Q8-AnyCPU" Version="7.14.4" />
+    <PackageReference Include="NetVips" Version="1.1.0" />
     <PackageReference Include="PhotoSauce.MagicScaler" Version="0.9.1" />
     <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-dev002604" />
     <PackageReference Include="SkiaSharp" Version="1.68.0" />
-    <PackageReference Include="System.Drawing.Common" Version="4.5.1" />
+    <PackageReference Include="System.Drawing.Common" Version="4.6.0" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(BenchmarkWithNuGetBinaries)' == 'true'">
-    <PackageReference Include="NetVips.Native" Version="8.8.0-rc2" />
+    <PackageReference Include="NetVips.Native" Version="8.8.3" />
     <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="1.68.0" />
   </ItemGroup>