Just saw this was updated and noticed a couple of issues with the benchmarks:

1) The MagicScaler method in the parallel test was actually calling the ImageMagick implementation, making it appear ~7x slower.

2) The new NetVips resize benchmark was effectively a no-op, giving it impossibly low timings.  Because libvips is a streaming/on-demand library, no work is done until pixels are materialized.  Both the original and resized "images" in that benchmark represented the *possibility* of pixels, not any actual work.  Materializing both the original blank image and the resized image by calling `CopyMemory()` brings the benchmark more in line with the others and with NetVips' performance in the real-world (load/resize/save) tests.

**Before**

```
|                    Method |         Mean |         Error |        StdDev | Ratio | RatioSD |    Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------------- |-------------:|--------------:|--------------:|------:|--------:|---------:|-------:|------:|----------:|
|   'System.Drawing Resize' | 10,959.96 us |    142.284 us |    36.9506 us | 1.000 |    0.00 |        - |      - |     - |     136 B |
|       'ImageSharp Resize' |  5,932.64 us |     29.762 us |     7.7291 us | 0.541 |    0.00 |        - |      - |     - |    7440 B |
|      'ImageMagick Resize' | 51,744.23 us | 31,533.143 us | 8,189.0570 us | 4.721 |    0.74 |        - |      - |     - |    5656 B |
|        'FreeImage Resize' |  8,022.36 us |    100.563 us |    26.1159 us | 0.732 |    0.00 | 500.0000 |      - |     - |     136 B |
|      'MagicScaler Resize' |  3,599.67 us |     44.185 us |    11.4747 us | 0.328 |    0.00 |  11.7188 |      - |     - |   59525 B |
| 'SkiaSharp Canvas Resize' |  4,775.94 us |     77.787 us |    20.2010 us | 0.436 |    0.00 |        - |      - |     - |     653 B |
| 'SkiaSharp Bitmap Resize' |  4,957.65 us |  1,535.034 us |   398.6435 us | 0.452 |    0.04 |        - |      - |     - |     848 B |
|          'NetVips Resize' |     42.74 us |      1.603 us |     0.4163 us | 0.004 |    0.00 |   0.9155 | 0.4272 |     - |    5080 B |
```

**After**

```
|                    Method |      Mean |     Error |    StdDev | Ratio | RatioSD |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |----------:|----------:|----------:|------:|--------:|---------:|------:|------:|----------:|
|   'System.Drawing Resize' | 10.947 ms | 0.1486 ms | 0.0386 ms |  1.00 |    0.00 |        - |     - |     - |     136 B |
|       'ImageSharp Resize' |  6.086 ms | 0.2766 ms | 0.0718 ms |  0.56 |    0.01 |        - |     - |     - |    7440 B |
|      'ImageMagick Resize' | 47.676 ms | 0.6292 ms | 0.1634 ms |  4.36 |    0.02 |        - |     - |     - |    5656 B |
|        'FreeImage Resize' |  8.105 ms | 0.2517 ms | 0.0654 ms |  0.74 |    0.01 | 500.0000 |     - |     - |     136 B |
|      'MagicScaler Resize' |  3.579 ms | 0.0704 ms | 0.0183 ms |  0.33 |    0.00 |  11.7188 |     - |     - |   59525 B |
| 'SkiaSharp Canvas Resize' |  4.735 ms | 0.0373 ms | 0.0097 ms |  0.43 |    0.00 |        - |     - |     - |     650 B |
| 'SkiaSharp Bitmap Resize' |  4.754 ms | 0.0484 ms | 0.0126 ms |  0.43 |    0.00 |        - |     - |     - |     848 B |
|          'NetVips Resize' |  5.246 ms | 0.1741 ms | 0.0452 ms |  0.48 |    0.00 |        - |     - |     - |    5320 B |
```

And finally, I updated the referenced imaging libraries to their current versions and updated the `TargetFrameworks` of the project to include both LTS (2.1) and Current (3.0) framework versions.

/cc @kleisauke